@@ -2,9 +2,9 @@ package ir.rahbod.habibi.adapter;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.constraint.ConstraintLayout;
-import android.support.v7.widget.RecyclerView;
-import android.util.Log;
+import androidx.constraintlayout.widget.ConstraintLayout;
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -15,7 +15,6 @@ import java.util.List;
 import ir.rahbod.habibi.R;
 import ir.rahbod.habibi.helper.PutKey;
 import ir.rahbod.habibi.helper.SessionManager;
-import ir.rahbod.habibi.model.Factor;
 import ir.rahbod.habibi.model.SubServiceItem;
 import ir.rahbod.habibi.pages.BottomSheetMain;
 import ir.rahbod.habibi.pages.RequestStepOneActivity;
