@@ -2,20 +2,18 @@ package ir.rahbod.habibi.adapter;
 
 import android.app.Dialog;
 import android.content.Context;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
-import android.widget.RelativeLayout;
 import android.widget.TextView;
 
 import java.util.List;
 
 import ir.rahbod.habibi.R;
-import ir.rahbod.habibi.model.Bill;
 import ir.rahbod.habibi.model.Transaction;
 
 public class AdapterTransaction extends RecyclerView.Adapter<AdapterTransaction.listViewHolder> {
