@@ -2,7 +2,7 @@ package ir.rahbod.habibi.pages;
 
 import android.app.Activity;
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.text.Editable;
 import android.text.TextWatcher;
@@ -12,8 +12,6 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ScrollView;
 import android.widget.Toast;
-
-import com.google.firebase.messaging.FirebaseMessaging;
 
 import java.io.IOException;
 
