@@ -1,8 +1,8 @@
 package ir.rahbod.habibi.pages;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.CardView;
+import androidx.cardview.widget.CardView;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.ImageView;
