@@ -1,15 +1,12 @@
 package ir.rahbod.habibi.pages;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import com.squareup.picasso.Picasso;
-
-import java.io.IOException;
 
 import ir.rahbod.habibi.R;
 import ir.rahbod.habibi.api.ApiClient;
