@@ -1,18 +1,12 @@
 package ir.rahbod.habibi.pages;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
 
 import ir.rahbod.habibi.R;
-import ir.rahbod.habibi.api.ApiClient;
-import ir.rahbod.habibi.api.ApiService;
-import ir.rahbod.habibi.model.TransactionList;
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
 
 public class TransactionActivity extends AppCompatActivity implements View.OnClickListener {
 
