@@ -3,7 +3,7 @@ package ir.rahbod.habibi.widget;
 import android.content.Context;
 import android.graphics.Canvas;
 import android.graphics.Typeface;
-import android.support.v7.widget.AppCompatTextView;
+import androidx.appcompat.widget.AppCompatTextView;
 import android.util.AttributeSet;
 
 public class FontTextViewBold extends AppCompatTextView {
