@@ -2,15 +2,13 @@ package ir.rahbod.habibi.controller;
 
 import android.app.Application;
 import android.content.Context;
-import com.crashlytics.android.Crashlytics;
-import io.fabric.sdk.android.Fabric;
 
 public class MyApp extends Application {
     public static volatile Context context;
     @Override
     public void onCreate() {
         super.onCreate();
-        Fabric.with(this, new Crashlytics());
+        //Fabric.with(this, new Crashlytics());
         context = getApplicationContext();
     }
 }
